+++ examples/example.rb
@@ -1,21 +1,21 @@
 require_relative '../lib/ruby_http_client'
 
-SendGrid::Config.new
+# This example uses the SendGrid API as an example
 headers = JSON.parse('
   {
     "Authorization": "Bearer ' + ENV['SENDGRID_API_KEY'] + '",
     "Content-Type": "application/json"
   }
 ')
-host = ENV['LOCAL_HOST']
+host =  'https://api.sendgrid.com'
 client = SendGrid::Client.new(host: host, request_headers: headers)
 
 # GET Collection
 query_params = { 'limit' => 100, 'offset' => 0 }
 response = client.version('v3').api_keys.get(query_params: query_params)
 puts response.status_code
-puts response.response_body
-puts response.response_headers
+puts response.body
+puts response.headers
 
 # POST
 request_body = JSON.parse('
@@ -30,15 +30,15 @@ request_body = JSON.parse('
 ')
 response = client.version('v3').api_keys.post(request_body: request_body)
 puts response.status_code
-puts response.response_body
-puts response.response_headers
-api_key_id = JSON.parse(response.response_body)['api_key_id']
+puts response.body
+puts response.headers
+api_key_id = JSON.parse(response.body)['api_key_id']
 
 # GET Single
 response = client.version('v3').api_keys._(api_key_id).get
 puts response.status_code
-puts response.response_body
-puts response.response_headers
+puts response.body
+puts response.headers
 
 # PATCH
 request_body = JSON.parse('
@@ -48,8 +48,8 @@ request_body = JSON.parse('
 ')
 response = client.api_keys._(api_key_id).patch(request_body: request_body)
 puts response.status_code
-puts response.response_body
-puts response.response_headers
+puts response.body
+puts response.headers
 
 # PUT
 request_body = JSON.parse('
@@ -63,10 +63,10 @@ request_body = JSON.parse('
 ')
 response = client.api_keys._(api_key_id).put(request_body: request_body)
 puts response.status_code
-puts response.response_body
-puts response.response_headers
+puts response.body
+puts response.headers
 
 # DELETE
 response = client.api_keys._(api_key_id).delete
 puts response.status_code
-puts response.response_headers
+puts response.headers