+++ lib/ruby_http_client.rb
@@ -4,27 +4,16 @@ module SendGrid
   require 'net/http'
   require 'net/https'
 
-  # Loads environment variables from a .env file.
-  class Config
-    def initialize
-        File.open('./.env').readlines.each do |line|
-        key, value = line.split '='
-        ENV[key] = value.chomp
-        end
-        ENV
-    end
-  end
-  
   # Holds the response from an API call.
   class Response
     # * *Args*    :
     #   - +response+ -> A NET::HTTP response object
     #
-    attr_reader :status_code, :response_body, :response_headers
+    attr_reader :status_code, :body, :headers
     def initialize(response)
       @status_code = response.code
-      @response_body = response.body
-      @response_headers = response.to_hash.inspect
+      @body = response.body
+      @headers = response.to_hash.inspect
     end
   end