+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2016-06-03
+### Changed
+- Made the Response variables non-redundant. e.g. response.response_body becomes response.body
+
+### Removed
+- Config class
+
 ## [1.1.0] - 2016-03-17
 ### Added
 - Config class moved to client